table: break manifest structs info separate header and compilation unit
