+++ configure.py
@@ -1163,6 +1163,7 @@ scylla_core = (['message/messaging_service.cc',
                 'db/size_estimates_virtual_reader.cc',
                 'db/snapshot-ctl.cc',
                 'db/snapshot/backup_task.cc',
+                'db/snapshot/manifest.cc',
                 'db/system_distributed_keyspace.cc',
                 'db/system_keyspace.cc',
                 'db/tags/utils.cc',
