+++ db/CMakeLists.txt
@@ -41,6 +41,7 @@ target_sources(db
     view/view_building_task_mutation_builder.cc
     snapshot-ctl.cc
     snapshot/backup_task.cc
+    snapshot/manifest.cc
     rate_limiter.cc
     per_partition_rate_limit_options.cc
     row_cache.cc
